+++ src/app/app.html
@@ -7,7 +7,6 @@
 }, $ctrl.language]">
   <header></header>
   <ui-view></ui-view>
-  <bungie-alerts></bungie-alerts>
   <toaster-container toaster-options="{ 'time-out': 10000, limit: 6 }"></toaster-container>
   <ng-include ng-if="::$ctrl.featureFlags.colorA11y" src="'data/color-a11y.svg'"></ng-include>
 </div>
