+++ src/app/whats-new/BungieAlerts.tsx
@@ -1,14 +1,29 @@
-import * as React from 'react';
-import { alerts$ } from '../shell/bungie-alerts.component';
-import { GlobalAlert } from '../bungie-api/bungie-core-api';
-import { Subscription } from 'rxjs/Subscription';
+import { equals } from 'angular';
 import { t } from 'i18next';
+import * as React from 'react';
+import { Observable } from 'rxjs/Observable';
+import { Subscription } from 'rxjs/Subscription';
+import { getGlobalAlerts, GlobalAlert } from '../bungie-api/bungie-core-api';
+import '../rx-operators';
 import './BungieAlerts.scss';
+
+// TODO: Move this back to a shared file so we can observe it from the header?
+const alerts$ = Observable.timer(0, 10 * 60 * 1000)
+  // Fetch global alerts, but swallow errors
+  .switchMap(() => Observable.fromPromise(getGlobalAlerts()).catch(() => Observable.empty<GlobalAlert[]>()))
+  .startWith([] as GlobalAlert[])
+  // Deep equals
+  .distinctUntilChanged<GlobalAlert[]>(equals);
+  // TODO: Publish? this is the part I never get
+  // TODO: redux?
 
 interface State {
   alerts: GlobalAlert[];
 }
 
+/**
+ * Displays maintenance alerts from Bungie.net.
+ */
 export default class BungieAlerts extends React.Component<{}, State> {
   private subscription: Subscription;
